+++ app/build.gradle
@@ -67,8 +67,8 @@ dependencies {
     implementation 'com.paypal.sdk:paypal-android-sdk:2.15.3'
     implementation 'androidx.navigation:navigation-fragment:2.3.4'
     implementation 'androidx.navigation:navigation-ui:2.3.4'
-    implementation 'androidx.lifecycle:lifecycle-livedata-ktx:2.3.0'
-    implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.3.0'
+    implementation 'androidx.lifecycle:lifecycle-livedata-ktx:2.3.1'
+    implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.3.1'
     implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
     implementation 'com.github.addisonelliott:SegmentedButton:3.1.9'
     implementation 'com.skyfishjy.ripplebackground:library:1.0.1'
