--- app/src/main/res/layout/activity_splashscreen.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:orientation="vertical"
-    android:layout_height="match_parent"
-    android:background="@color/AlikappGris"
-    tools:context=".splashScreen"
-    >
-
-    <ImageView
-        android:id="@+id/imagenlogo"
-        android:layout_width="match_parent"
-        android:layout_height="200dp"
-        android:layout_marginTop="200dp"
-        android:foregroundGravity="center"
-        android:src="@mipmap/logo_alikapp"
-        tools:layout_editor_absoluteX="105dp"
-        tools:layout_editor_absoluteY="265dp" />
-
-    <TextView
-        android:id="@+id/textAlikapp"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:fontFamily="@font/abeezee"
-        android:gravity="center"
-        android:text=" Urgencias Mecanicas"
-        android:textColor="@color/AlikappBlue"
-        android:textSize="18sp"
-        android:layout_marginStart="-20dp"
-        tools:ignore="MissingConstraints"
-        tools:layout_editor_absoluteX="139dp"
-        tools:layout_editor_absoluteY="696dp" />
-
-
-</LinearLayout>
